DSAD/requirements.txt
scapy>=2.5.0
flask>=2.0.0
flask-cors>=3.0.10
python-dotenv>=1.0.0
netifaces>=0.11.0
